backend/requirements.txt
# FastAPI Framework
fastapi==0.95.0
uvicorn[standard]==0.22.0

# MongoDB
pymongo==4.3.3
motor==3.1.1

# Validación
pydantic==1.10.7
pydantic-settings==2.0.1

# CORS y seguridad
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilidades
python-dotenv==1.0.0
requests==2.28.2

# Opcional: Documentación automática (ya incluido en FastAPI)
# docs adicionales

